QAA project
Our client is a highly-respected large-cap company that has a significant sum of assets in its corporate pension, and would like one component of its pension assets to be invested in a fixed-maturity multi-asset scheme with a regular payout.

The main characteristics of this special portion of pension are the following:

All assets will be given back to the retirees in 10 years
The initial investment is €120M, and each month, the payout should be MAX(€1.001M,P), where P is the 1/n fraction of total remaining assets each month, where n is the number of remaining months to the maturity (i.e. 1/120 for the first payout and 1/119 for the second etc)
It is in the retiree’s interest to maximize P
If the remaining total assets become less than €1.001M, however, this company is liable to meet the shortfall
Your report should include the elements listed below:

What kind of investment strategy would you propose to our client?
Why would our client implement your strategy?
How likely would your strategy lead our client to the case of meeting shortfalls?
Support your strategy by using historical data
Highlights:
Our client is a highly-respected large-cap company that has a significant sum of assets
in their corporate pension, and would like one component of their pension assets to be
invested in a fixed-maturity multi-asset scheme with a regular payout.